BlackBox/requirements.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
diffusers==0.29.2
transformers==4.40.1
accelerate==0.28.0
safetensors==0.4.2
huggingface_hub==0.23.4
datasets==2.19.1
Pillow==10.3.0
scikit-image==0.22.0
scipy==1.12.0
opencv-python==4.9.0.80
lpips==0.1.4
scikit-learn==1.4.2
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.66.2
pandas==2.2.2
numpy==1.26.4
